+++ Changelog.md
@@ -8,8 +8,9 @@ Coming Soon
 - More built-in encoding systems (base64 & sha-1)
 - Multipart request parsing
 - CookieJar struct for holding Cookies in the request
-- RealIp extention
-- Allow serving an IPv6 addrs
+- RealIp extension
+- Allow serving an IPv6 addr
+- Use a `Headers` struct to hold default headers
 
 # 2.0.0